+++ CHANGELOG.md
@@ -2,7 +2,8 @@ Elgg Cropper - CHANGE LOG
 ========================
 
 ### 2.1.1 (2016-11-15)
-* **releases:** Plenty of Cropper tool options are now available on plugin settings
+* Plenty of Cropper tool options are now available on plugin settings
+* Added option to select size of preview box in settings
 
 ### 2.1.0 (2016-11-07)
 * **releases:** Initial release